feat: move query and extent to own files
