+++ mapproxy/seed/util.py
@@ -23,7 +23,7 @@ import time
 from datetime import datetime
 import pickle
 
-from mapproxy.layer import map_extent_from_grid
+from mapproxy.extent import map_extent_from_grid
 from mapproxy.util.fs import write_atomic
 
 import logging
