+++ mapproxy/source/ogcapitiles.py
@@ -22,12 +22,8 @@ from threading import Lock
 from mapproxy.client.http import HTTPClientError
 from mapproxy.grid.tile_grid import tile_grid_from_ogc_tile_matrix_set
 from mapproxy.image.opts import ImageOptions
-from mapproxy.layer import (
-    BlankImage,
-    MapExtent,
-    DefaultMapExtent,
-    MapLayer,
-)
+from mapproxy.layer import BlankImage, MapLayer
+from mapproxy.extent import MapExtent, DefaultMapExtent
 from mapproxy.source import SourceError, InvalidSourceQuery
 from mapproxy.srs import ogc_crs_url_to_auth_code
 from mapproxy.util.py import reraise_exception
