+++ mapproxy/test/unit/test_cache.py
@@ -44,12 +44,12 @@ from mapproxy.layer import (
     CacheMapLayer,
     DirectMapLayer,
     MapBBOXError,
-    MapExtent,
     MapLayer,
-    MapQuery,
     ResolutionConditional,
     SRSConditional,
 )
+from mapproxy.extent import MapExtent
+from mapproxy.query import MapQuery
 from mapproxy.request.wms import WMS111MapRequest
 from mapproxy.source import InvalidSourceQuery, SourceError
 from mapproxy.source.tile import TiledSource
