+++ mapproxy/test/unit/test_cache_geopackage.py
@@ -25,7 +25,7 @@ from mapproxy.cache.geopackage import GeopackageCache, GeopackageLevelCache
 from mapproxy.cache.tile import Tile
 from mapproxy.grid.tile_grid import tile_grid, TileGrid
 from mapproxy.image import ImageSource
-from mapproxy.layer import MapExtent
+from mapproxy.extent import MapExtent
 from mapproxy.srs import SRS
 from mapproxy.test.helper import assert_files_in_dir, assert_permissions
 from mapproxy.test.unit.test_cache_tile import TileCacheTestBase
