+++ mapproxy/test/unit/test_conf_loader.py
@@ -29,7 +29,7 @@ from mapproxy.config.configuration.base import ConfigurationError
 from mapproxy.config.configuration.proxy import ProxyConfiguration
 from mapproxy.cache.tile import TileManager
 from mapproxy.config.spec import validate_options
-from mapproxy.layer import MapExtent
+from mapproxy.extent import MapExtent
 from mapproxy.seed.spec import validate_seed_conf
 from mapproxy.srs import SRS
 from mapproxy.test.helper import TempFile
